requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
attrs==24.2.0
backoff==2.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.5
fonttools==4.55.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
inference-sdk==0.27.0
jiter==0.7.1
kiwisolver==1.4.7
marshmallow==3.23.1
matplotlib==3.9.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.54.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==10.4.0
propcache==0.2.0
py-cpuinfo==9.0.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
requests==2.32.3
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
supervision==0.22.0
tqdm==4.67.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
yarl==1.17.1
